fastapi-services/Databases/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
beautifulsoup4==4.13.4
boto3==1.40.60
botocore==1.40.60
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorclass==2.2.2
compressed-rtf==1.0.7
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
easygui==0.98.3
ebcdic==1.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
extract-msg==0.54.1
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
git-filter-repo==2.47.0
gunicorn==23.0.0
gymnasium==1.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
lark==1.1.9
lxml==5.4.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.4.2
mysql-connector-python==9.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
odfpy==1.4.1
olefile==0.47
oletools==0.60.2
openai==1.84.0
openpyxl==3.1.5
oracledb==3.2.0
packaging==25.0
pandas==2.3.1
pcodedmp==1.2.6
pdfkit==1.0.0
pillow==11.3.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyodbc==5.2.0
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-nmap==0.7.1
python-pptx==1.0.2
pytz==2025.2
red-black-tree-mod==1.22
redis==6.4.0
requests==2.32.4
RTFDE==0.1.2.1
s3transfer==0.14.0
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stable_baselines3==2.7.0
starlette==0.46.2
striprtf==0.0.29
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.34.2
xlrd==2.0.2
xlsxwriter==3.2.5
